+++ src/components/links/link/Link.svelte
@@ -7,9 +7,17 @@
     type IconLocation,
     iconLocationDefault,
   } from "./types/LinkIconLocation";
-  import openInNewWindow from "../../../common/utils/openInNewWindow";
   import { linkPriorityDefault, type LinkPriority } from "./types/LinkPriority";
   import type { IconOpacity } from "../../../icons/types/IconOpacity";
+
+  const openInNewWindow = (
+    url: string,
+    width: number = 700,
+    height: number = 650
+  ) => {
+    const windowFeatures = `left=100,top=100,width=${width},height=${height}}`;
+    window.open(url, "new", windowFeatures);
+  };
 
   const onClick = (event: MouseEvent) => {
     if (href !== "" && opensNewWindow === "true") {